fix(tests): TestGGUFMetaArrayStaging passes an Int64 array to AddMetaInt32Array
The staging fixture declared its token_type array as `array of integer`,
which does not match AddMetaInt32Array's `array of Int64` parameter, so
the suite did not compile.

+++ tests/TestNeuralPretrained.pas
@@ -6953,7 +6953,7 @@ var
   GGUFPath: string;
   Tokens: array of string;
   Scores: array of single;
-  Types: array of integer;
+  Types: array of Int64;
   Dummy: TNNetVolume;
   SavedStage, StageCnt, i: integer;
 begin
